small fixes

+++ 1/1_7/main.c
@@ -2,7 +2,7 @@
 
 
 int main(int argc, char* argv[]) {
-	for (size_t j = 1;j < argc; ++j) {
+	for (int j = 1;j < argc; ++j) {
 		char * catalog_name = argv[j];
 		error_msg errorMsg = processing_catalog(catalog_name);
 		if(errorMsg.type){
